+++ readme_improver/__init__.py
@@ -1,0 +1,1 @@
+"Utilities for the ai-readme-improver package."
