+++ readme_improver/readme_loader.py
@@ -1,4 +1,4 @@
-from logger import get_logger
+from .logger import get_logger
 
 
 logger = get_logger()
