+++ tests/test_improver.py
@@ -1,4 +1,8 @@
-from improver import generate_summary, suggest_improvements, rewrite_readme
+from readme_improver.improver import (
+    generate_summary,
+    suggest_improvements,
+    rewrite_readme,
+)
 
 
 def dummy_ask(prompt, model="gpt-3.5-turbo", temperature=0.5, max_tokens=800):
@@ -6,18 +10,18 @@ def dummy_ask(prompt, model="gpt-3.5-turbo", temperature=0.5, max_tokens=800):
 
 
 def test_generate_summary(monkeypatch):
-    monkeypatch.setattr("improver.ask_openai", dummy_ask)
+    monkeypatch.setattr("readme_improver.improver.ask_openai", dummy_ask)
     result = generate_summary("Readme")
     assert "RESPONSE" in result
 
 
 def test_suggest_improvements(monkeypatch):
-    monkeypatch.setattr("improver.ask_openai", dummy_ask)
+    monkeypatch.setattr("readme_improver.improver.ask_openai", dummy_ask)
     result = suggest_improvements("Readme")
     assert "RESPONSE" in result
 
 
 def test_rewrite_readme(monkeypatch):
-    monkeypatch.setattr("improver.ask_openai", dummy_ask)
+    monkeypatch.setattr("readme_improver.improver.ask_openai", dummy_ask)
     result = rewrite_readme("Readme")
     assert "RESPONSE" in result
